chore: rename salesaya-constant to constant
